% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Randomforest_fcts.R
\name{cat.distrib.plot}
\alias{cat.distrib.plot}
\title{Function to plot violin plots of variables importance based on their category
and return one plot with statistics and Kruskall Wallis test}
\usage{
cat.distrib.plot(rf_df, metric_nm, palette, drivers_nm_df)
}
\arguments{
\item{rf_df}{the data frame gathering random forests results from the
\code{test.rf.model} function.}

\item{metric_nm}{character string referring to the studied metric}

\item{palette}{a color palette containing the color names of the broad
categories of the drivers present in the \code{taxa_plot_df}}

\item{drivers_nm_df}{a data frame containing shortened names of drivers}
}
\value{
two plots in a list, one with stats and mean and the other without
and print the result of a Kruskall-Wallis test to see if there is a
significant difference of median variables importance between drivers categories
}
\description{
Function to plot violin plots of variables importance based on their category
and return one plot with statistics and Kruskall Wallis test
}
