% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute_missing_traits_fcts.R
\name{check.genus}
\alias{check.genus}
\title{Function to complete the taxonomic data frame if data already
gathered on missing Genus}
\usage{
check.genus(taxo_df, missing_sp_vect)
}
\arguments{
\item{taxo_df}{the data frame output of the \code{taxize::tax_name()} function}

\item{missing_sp_vect}{a vector containing the name of species with missing
taxonomic data}
}
\value{
the taxonoic dataframe completed where it can be.
}
\description{
Function to complete the taxonomic data frame if data already
gathered on missing Genus
}
