% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_traits_fcts.R
\name{check.quantiles}
\alias{check.quantiles}
\title{Get names of species which have extreme traits values}
\usage{
check.quantiles(traits_df, quant_traits_vect)
}
\arguments{
\item{traits_df}{species*traits data frame with species as a column}

\item{quant_traits}{a vector containing the name of quantitative traits}
}
\description{
Get names of species which have extreme traits values
}
