% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Randomforest_fcts.R
\name{circular.drivers.plot}
\alias{circular.drivers.plot}
\title{Plot n drivers for a given taxa and all diversity metrics}
\usage{
circular.drivers.plot(taxa_plot_df, drivers_nm_df, palette, div_facet)
}
\arguments{
\item{taxa_plot_df}{data frame from the \code{create.df.circular.plot} function}

\item{drivers_nm_df}{a data frame containing shortened names of drivers}

\item{palette}{a color palette containing the color names of the category present
in the \code{taxa_plot_df}}

\item{div_facet}{a character string referring to whether FD of PD is studied here.
It could be either \code{"FD"} or \code{"PD"}.}
}
\description{
Plot n drivers for a given taxa and all diversity metrics
}
