% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Null_models_fcts.R
\name{compute.null.model.metrics}
\alias{compute.null.model.metrics}
\title{Function to compute SES and Realised/Potential ratios based on null models}
\usage{
compute.null.model.metrics(
  null_model_df,
  null_metric_to_compute,
  ind_values_df
)
}
\arguments{
\item{null_model_df}{the data frame gathering for each  grid cell the
simulated values of a given metric. Grid cells are in row and null asb in
columns. It's the output of the \code{compute.null.models.FD} or
\code{compute.null.models.PD} functions.}

\item{null_metric_to_compute}{a vector gathering character strings referring to
the metric to compute. Can be either \code{c("ses")}, \code{c("R/P")} or
\code{c("ses", "RP_ratio")}.}

\item{ind_values_df}{a data frame gathering the real values of the PD or
FD metric computed.}
}
\value{
a list with one or two data frames (according to the number of
metrics to compute) gathering for each grid cell (rows), the metric value
(column).
}
\description{
Function to compute SES and Realised/Potential ratios based on null models
}
