% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_div_per_cell_fcts.R
\name{compute.PD.per.cell}
\alias{compute.PD.per.cell}
\title{Compute phylogenetic indices}
\usage{
compute.PD.per.cell(sp_asb_df, phylo, metric_nm, grid, taxon)
}
\arguments{
\item{sp_asb_df}{a dataframe linking species (columns) and cells (rows) for
a given scale}

\item{phylo}{a phylogeny as a phylo object with only species present in the
\code{sp_asb_df}}

\item{metric_nm}{a character string refering to the name of the metric to be
computed, could be either: "Faith_index", "MPD", "meanED".
Faith index is computed through the \code{picante::pd()} function. MPD
metric is computed throught the \code{picante::mpd()} , mean(ED) is computed
through the \code{picante::evol.distinct} as the mean of ED for the studied
subset of species (ED computed on the global tree).}

\item{grid}{a character string referring to the scale used. Can either be
"50x50" or "10x10".}

\item{taxon}{a character string referring to the taxa studied. Can either be
"Trees", "Birds" or "Butterflies".}
}
\value{
a data frae with cells as rows and the following columns:
\code{Idgrid}, \code{metric}, \code{Grid}, \code{Taxon}. Can be used to plot
the metric thorugh the \code{div.maps.plot} function
}
\description{
Compute phylogenetic indices
}
