% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_div_per_cell_fcts.R
\name{compute.TD.per.cell}
\alias{compute.TD.per.cell}
\title{Compute Taxonomic Diversity At A Given Scale}
\usage{
compute.TD.per.cell(occ_df, scale)
}
\arguments{
\item{occ_df}{a data frame with the spatial cells in rows and the following
columns: \strong{Taxon} with Taxon name (should be the same for all df as
one taxon is studied here), \strong{Species} with species name
(Genus_species_subspecies), \strong{Region} with Region's name (should be the
same for all df as one area studied here), \strong{Grid} with either
\code{50x50} or \code{10x10} values as they are the two different scales
studied and \code{Idgrid} with the identity of each cell (Note: 50x50km and
10x10km cells have different format of ID).}

\item{scale}{a character string referring to the scale at which the analysis
should be realised. Could be either \code{10x10} or \code{50x50}.}
}
\value{
a data frame with the values of taxonomic diversity (here species
richness) per cell of the chosen grid (Chosen grid reminded as a column),
with a column referring to the Taxon studied (should be the same for all df as
one taxon is studied here)
}
\description{
Compute Taxonomic Diversity At A Given Scale
}
