% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_metrics_landuse_fcts.R
\name{compute.wm.wsd}
\alias{compute.wm.wsd}
\title{Compute weighted mean and sd across different resolutions}
\usage{
compute.wm.wsd(landuse_df)
}
\arguments{
\item{landuse_df}{a dataframe with the INTEGRADIV_en db columns,
reducted to the 50*50 scale}
}
\value{
two dataframes (one for past, one for present) with the following
columns \code{IdGrid}, \code{Type}, \code{Variable_Code}, \code{Metric} and
\code{Value} were value reflect the weighted mean and weighted sd of the
\% of the cell filled by the studied land use
}
\description{
Compute weighted mean and sd across different resolutions
}
