% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_plot_metrics_correlation_fcts.R
\name{correlation.fd.pd}
\alias{correlation.fd.pd}
\title{Compute correlation between PD and FD metrics and plot}
\usage{
correlation.fd.pd(indices_df, save)
}
\arguments{
\item{indices_df}{a data frame gathering the indices to compare: they should
be fric, fdis, fspe, mpd,faith pd and species richness and also
colmns for \code{Idgrid}, \code{Taxon} and \code{Scale}. The df has been
created here under \code{5_..._Correlation_btw_..._metrics.R}. Should be in
the following order and names: \code{Idgrid, sp_richn, fric, fdis, fspe,
faith_pd, mpd, Grid, Taxon}.}
}
\value{
two ggplot plot showing correlation between all variables and species
richness as dot size (for the second plot)
}
\description{
Compute correlation between PD and FD metrics and plot
}
