% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_traits_fcts.R
\name{cramerv.compute}
\alias{cramerv.compute}
\title{Compute Cramer's V on a set of traits}
\usage{
cramerv.compute(sp_tr_df, var_nm)
}
\arguments{
\item{sp_tr_df}{dataframe gathering species in rows and traits in columns}

\item{var_nm}{a vector gathering the name of the traits to study (so only
categorical ones)}
}
\value{
a dataframe with Cramer's V values for eauch pair of trait
}
\description{
Compute Cramer's V on a set of traits
}
