% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Randomforest_fcts.R
\name{create.df.circular.plot}
\alias{create.df.circular.plot}
\title{Create a data frame to plot the circular plot of drivers for each taxa}
\usage{
create.df.circular.plot(rf_df_list, var_nb)
}
\arguments{
\item{rf_df_list}{a list containing the rf results from the
\code{test.rf.model} function - each element should be given the name of the
metric studied.}

\item{var_nb}{a number referring to the number of variables that should be
kept in the graph - variables ordered according to their importance.}
}
\value{
a data frame with the following columns: Driver_nm, Div_metric,
Driver_imp, Driv_cat
}
\description{
Create a data frame to plot the circular plot of drivers for each taxa
}
