% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_div_maps_fcts.R
\name{div.maps.plot}
\alias{div.maps.plot}
\title{Plot A Given Diversity Metric At A Given Scale}
\usage{
div.maps.plot(
  div_per_cell_df,
  div_facet_nm,
  metric_nm,
  grid,
  continuous,
  col_pal,
  plot_title,
  save
)
}
\arguments{
\item{div_per_cell_df}{a data frame with the spatial cells in rows and the
following columns: \strong{Idgrid} with the identity of each cell,
\strong{metric} name of the metric for column name (ex: \code{sp_richness}),
\strong{Grid} referring to the scale studied (only one value here),
\strong{Taxon} with Taxon's name (should be the same for all df as
one taxon is studied here). Columns should be \bold{in this order}.}

\item{div_facet_nm}{a character string referring to the name of the diversity
facets studied: Can be either \code{TD}, \code{FD} or \code{PD}.}

\item{metric_nm}{a character string referring to the name of the diversity
metric studied.}

\item{grid}{a \code{sf} object referring to the spatial grid used, either
the 10x10km grid (WOODIV one) or the 50x50km one (INTEGRADIV one).}

\item{continuous}{if continuous equals \code{TRUE}, diversity values are
plotted along a continuous scale. Otherwise they are plotted along a
discrete scale defined by 1/10 of values.}

\item{plot_title}{a TRUE/FALSE value referring to plotting or not the graph
title (on the diversity facet studied)}

\item{save}{a TRUE/FALSE value referring to saving or not the graph}
}
\value{
a data frame with the values of taxonomic diversity (here species
richness) per cell of the chosen grid.
}
\description{
Plot A Given Diversity Metric At A Given Scale
}
