% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_metrics_landuse_fcts.R
\name{divide.pres.past}
\alias{divide.pres.past}
\title{Divide the dataset in present and past dataframes}
\usage{
divide.pres.past(landuse_df)
}
\arguments{
\item{landuse_df}{a dataframe with the INTEGRADIV_en db columns,
reducted to the 50*50 scale}
}
\value{
two dataframes (one for past, one for present) with the following
columns \code{IdGrid}, \code{Time} and \code{Value} where Time reflects the
time period for which the metric has been computed and value reflect the \% of
the cell filled by the studied land use
}
\description{
Divide the dataset in present and past dataframes
}
