% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Randomforest_fcts.R
\name{get.ale.data}
\alias{get.ale.data}
\title{Get data to plot ALE plots (one data for each 100 model)}
\usage{
get.ale.data(rf_mod, rf_data, chosen_var, ale_list, i)
}
\arguments{
\item{rf_mod}{a given random forest (one of the 100 run)}

\item{rf_data}{data used to run the random forest}

\item{chosen_var}{a vector containing the names of the variables
for which the ALE plot should be used}

\item{ale_list}{the list containing coordinates for ALE plots for
all variables (updated at each round of new rf model)}

\item{i}{the iteration number (nuber of the rf model to be studied
based on the 100 iterations)}
}
\description{
Get data to plot ALE plots (one data for each 100 model)
}
