% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Randomforest_fcts.R
\name{heatmap.varimp}
\alias{heatmap.varimp}
\title{Plot a heatmap illustrating the importance of each variable for the three taxa}
\usage{
heatmap.varimp(rf_all_taxa_list, metric_nm, plot_nb)
}
\arguments{
\item{rf_all_taxa_list}{a list containing the result of the
\code{test.rf.model} function for all the taxa studied.}

\item{metric_nm}{a vector containing the name of the studied metric}

\item{plot_nb}{a TRUE/FALSE value to indicate whether the mean\%IncMSE is to
be plotted or not on the heatmap}
}
\value{
a heatmap with species on columns and drivers as rows,
gathered by type and colors refelecting variable importance.
}
\description{
Plot a heatmap illustrating the importance of each variable for the three taxa
}
