% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute_missing_traits_fcts.R
\name{impute.missing.traits}
\alias{impute.missing.traits}
\title{Compute mean/mode at the closest taxonomic level with information to impute
missing values}
\usage{
impute.missing.traits(sp_tr_NA_df, taxo_df, traits_NA_nms)
}
\arguments{
\item{traits_NA_nms}{}
}
\description{
Compute mean/mode at the closest taxonomic level with information to impute
missing values
}
