% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Help_fcts.R
\name{locate.cells}
\alias{locate.cells}
\title{Localise cells in a given grid}
\usage{
locate.cells(cell_vect, grid)
}
\arguments{
\item{cell_vect}{a vector containing the name (Idgrid) of the cells to
localise.}

\item{grid}{a \code{sf} object referring to the spatial grid used, either
the 10x10km grid (WOODIV one) or the 50x50km one (INTEGRADIV one).}
}
\value{
a ggplot object plotting in red where the given cells are located
}
\description{
Localise cells in a given grid
}
