% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Randomforest_fcts.R
\name{plot.ale}
\alias{plot.ale}
\title{Plot ALE plots for all variables (mean and sd)}
\usage{
\method{plot}{ale}(ale_list, chosen_var, metric_nm, taxa_nm, drivers_nm_df)
}
\arguments{
\item{ale_list}{a list containing for each variable (driver), a dataframe
with coordinates of points of the ALE plot}

\item{chosen_var}{a vector containing the names of all the variables to study}

\item{metric_nm}{a character string refering to the metric studied}

\item{taxa_nm}{a character string refering to the taxa studied}

\item{drivers_nm_df}{a dataframe linking drivers raw names and nice names}
}
\description{
Plot ALE plots for all variables (mean and sd)
}
