% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Randomforest_fcts.R
\name{plot.partial.dependence}
\alias{plot.partial.dependence}
\title{Take a list of models, a predictor variable name, and the training data,
then computes and combines the partial dependence plots for the given
variable across all models.}
\usage{
\method{plot}{partial.dependence}(models, var_name, data, iteration_nb)
}
\arguments{
\item{data}{}
}
\description{
Take a list of models, a predictor variable name, and the training data,
then computes and combines the partial dependence plots for the given
variable across all models.
}
