% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Randomforest_fcts.R
\name{relationships.plot}
\alias{relationships.plot}
\title{Plot six groups of relationships between drivers and diversity - one group per
category of drivers}
\usage{
relationships.plot(ses_var_df, metric_nm, palette, drivers_nm_df)
}
\arguments{
\item{ses_var_df}{the data frame for each cell, values of each driver and
value of studied SES.}

\item{metric_nm}{character string referring to the studied metric}

\item{palette}{a color palette containing the color names of the broad
categories of the drivers present in the \code{taxa_plot_df}}

\item{drivers_nm_df}{a data frame containing shortened names of drivers}
}
\value{
eight ggplot objects in a list - in each object are represented the
relationships between diversity facets and specific drivers belonging to a
given category (but 2 plots for past land use and present human impact as
too many variables otherwise), each individual driver being represented as a
ggplot2 facet
}
\description{
Plot six groups of relationships between drivers and diversity - one group per
category of drivers
}
