% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Randomforest_fcts.R
\name{test.rf.model}
\alias{test.rf.model}
\title{Compute n iteration of the random forest model to test its accuracy and
variable importance (standardised and untandardised one) and compute
ALE Plots.}
\usage{
test.rf.model(rf_data, iteration_nb, metric_nm, taxa_nm, plot, drivers_nm_df)
}
\arguments{
\item{rf_data}{a data frame containing data to run the random forest - ie
predictor variables in columns with the last column being the ses of a given
diversity metric. This last column must be named \code{ses}.}

\item{iteration_nb}{the number of time the random forest should be repeated.}

\item{metric_nm}{a character string referring to the name of the metric studied
(used for saving partial dependance files)}

\item{taxa_nm}{a character string referring to the name of the taxa studied
(used for saving partial dependance files)}

\item{plot}{a TRUE/FALSE value according to whether or not partial dependance
are to be plotted and saved (it takes a lot of time)}

\item{drivers_nm_df}{a data frame containing shortened names of drivers}
}
\value{
a data frame with variable importance for each rf, another
dataframe in the same format with std importance (so that the most
important variable has a value of 1 and negative importances equal 0).
These two dataframes contains the mean and sd for each variable importance
across the n iterations of random forests. The function also returns the
mean and sd R squared (explained variance of the model computed on the
OOB data). Lastly, the funtion returns ALE plots for each driver studied.
}
\description{
Compute n iteration of the random forest model to test its accuracy and
variable importance (standardised and untandardised one) and compute
ALE Plots.
}
