% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Randomforest_fcts.R
\name{varimp.plot}
\alias{varimp.plot}
\title{Plot the variable importance for a given taxa and metric (lollipop plot)}
\usage{
varimp.plot(var_imp_df)
}
\arguments{
\item{var_imp_df}{output of the \code{test.rf.model} function.}
}
\value{
a lollipop plot with drivers on columns and mean \%IncMSE on x axis
with colors referring to drivers category
}
\description{
Plot the variable importance for a given taxa and metric (lollipop plot)
}
